pyproject.toml
[tool.poetry]
name = "rembish_org"
version = "0.4.0"
description = "Personal website"
authors = ["Alex Rembish <alex@rembish.org>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
uWSGI = "^2.0.19"
markdown2 = "^2.3.10"
Flask-WTF = "^0.14.3"
requests = "^2.24.0"
Flask-SQLAlchemy = "^2.4.4"
Flask-Security = "^3.0.0"
Flask-Dance = "^3.0.0"
email-validator = "^1.1.1"
SQLAlchemy-Utils = "^0.36.8"
PyMySQL = "^0.10.1"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.14.0"
pyOpenSSL = "^19.1.0"
Flask-DebugToolbar = "^0.11.0"
docker-compose = "^1.27.4"
semver = "^2.13.0"
toml = "^0.10.1"
Flask-Migrate = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
